[fix] 用户充值额度过高，数据中心推送科学计数法修正

+++ src/main/java/com/tianli/chain/dto/BaseTokenReq.java
@@ -26,7 +26,7 @@ public class BaseTokenReq {
 
     private String block;
 
-    private BigInteger value;
+    private BigDecimal value;
 
     private String contractAddress;
